Added stuff to the trap menu and switched the movement script
I made it so that now the trap menu pulls up with the tactical view and you can now spend points in order to get landmines. Made it so that it tells you the cost and the item manages to stay in the trap slot without moving around the screen. Also changed the movement script and connected it to the gun so that when your gun is drawn you move slower than when it is holstered.